Move github.com/cilium/hubble/pkg/testutils to local package and use it
Copy package github.com/cilium/hubble/pkg/testutils into the existing
internal package github.com/cilium/tetragon/pkg/testutils

Adjust local use of the package to the internal version to drop a
dependency on standalone Hubble package github.com/cilium/hubble
